+++ fancyflags/__init__.py
@@ -30,6 +30,7 @@ from fancyflags._definitions import define_flags
 
 # Automatically build fancyflags defs from a callable signature.
 from fancyflags._auto import auto
+from fancyflags._auto import auto_from_value
 from fancyflags._define_auto import DEFINE_auto
 
 # `Item` definitions for supported types.